pyproject.toml
[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.poetry]
name = "python-utility-scripts"
version = "0.0.1"
description = "Python utility scripts"
authors = ["dbasunag <dbasunag@redhat.com>"]
license = "Apache License 2.0"
readme = "README.md"
repository = "https://github.com/RedHatQE/python-utility-scripts"
homepage = "https://github.com/RedHatQE/python-utility-scripts"
documentation = "https://github.com/RedHatQE/python-utility-scripts/blob/main/README.md"
classifiers = [
"Programming Language :: Python :: 3",
"Operating System :: OS Independent",
]

packages = [
    { include = "apps" },
]

[tool.poetry.scripts]
pyappsutils-unusedcode = "apps.unused_code.unused_code:get_unused_functions"
pyappsutils-polarion_tc_requirements = "apps.polarion.polarion_verify_tc_requirements:has_verify"


[tool.poetry.dependencies]
python = "^3.8"
python-simple-logger = "^1.0.8"
pylero = "^0.0.8"
openshift-python-utilities = "^5.0.23"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "*"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"

[tool.poetry-dynamic-versioning.substitution]
files = ["VERSION"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
